+++ tests/PdoMariaDbStorage/StorageSelectTest.php
@@ -23,7 +23,7 @@ use PDO;
 /**
  * StorageSelectTest
  */
-class StorageSelectTest extends \Tobento\Service\Storage\Test\StorageSelectTest
+class StorageSelectTest extends \Tobento\Service\Storage\Test\StorageSelect
 {
     protected null|PdoDatabase $database = null;
     
@@ -62,8 +62,12 @@ class StorageSelectTest extends \Tobento\Service\Storage\Test\StorageSelectTest
         $this->dropTable($this->tableProductsLg);
     }
     
-    protected function dropTable(Table $table): void
+    protected function dropTable(null|Table $table): void
     {
+        if (is_null($table) || is_null($this->database)) {
+            return;
+        }
+        
         $table->dropTable();
         
         $processor = new PdoMySqlProcessor();
